feat(mail): Additional properties for SMTP configuration
Closes gravitee-io/issues#488
